Web Scrapping
Here are some of the python scripts that can be used to scrap data from any website. All you need is to replace link in the script for which you want to scrap data. Following scripts are added:

1- Scrap Links from a website 2- Scrap Titles, Headings, Text from a website 3- Scrap Images from the website 4- Scrap Specific data like Person Emails, Sales, Finanicials and addresses

Web scraping of company Profiles:

The project can Scrap 120 Millions of data from the website.

It scraps the following items from the website www.dnb.com

1- Company Name 2- Street Address 3- Regional Address 4- Phone 5- Website Link 6- Company Type 7- Number of Employees 8- Total Sales Revenue 9- Board Members 10- Founding year

The scraper saves data in CSV format. The scraper is dynamic and can be used for other links of the website.

#Python#BeautifulSoup#Requests#time#CSV

Contact me for more scraping scripts.